pyproject.toml
[tool.poetry]
name = "container-apps-event-driven-jobs-tutorial"
version = "0.1.0"
description = ""
authors = ["Yujin109 <koba.you.3939@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
typing-extensions = "^4.8.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
pandas = "^2.1.3"
openai = "^1.3.3"
requests = "^2.31.0"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
azure-storage-queue = "^12.9.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
black = "^23.10.1"
isort = "^5.12.0"
pyproject-flake8 = "^6.0.0.post1"
jupyter-client = "^8.3.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
flake8 = "^6.1.0"
typing-extensions = "^4.8.0"
types-requests = "^2.31.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.venv
  | \.git
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 119
extend-ignore = ["E203", "W503"]
exclude = [".venv", ".git", "__pycache__"]
max-complexity = 15

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true